uuIdの設定
